pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "redex"
version = "1.0.0"
description = "Redox potential prediction with graph neural networks and transfer learning"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "torch-geometric>=2.4",
    "rdkit>=2023.03",
    "scikit-learn>=1.3",
    "pandas>=2.0",
    "numpy>=1.24",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "matplotlib>=3.7",
    "scienceplots",
    "optuna>=3.0",
    "wandb",
]

[tool.setuptools.packages.find]
include = ["RedEx*"]
